fix: When we cast the "avg_value_m2" column to Polars, we do so to Float64 (recommended for Snowflake), instead of Float32.

+++ dags/fomento_valuations_ingestion_dag.py
@@ -89,7 +89,7 @@ def ingest_and_upload_fomento_valuations(s3_bucket: str, s3_key: str, aws_conn_i
         df_pl = (
             df_pl
             .with_columns([
-                pl.col("avg_value_m2").cast(pl.Float32, strict=False), 
+                pl.col("avg_value_m2").cast(pl.Float64, strict=False), 
                 pl.col("total_appraisals").cast(pl.Int32, strict=False), 
                 pl.lit(datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S")).alias("__LOADED_AT")
             ])
